// src/boilerplate/dataEngineApis/models/call-details-bean.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Data Engine
 * Collection of all Data Engine Requests which comprises of Reports, Stats, Monitoring and VLA Requests
 *
 * The version of the OpenAPI document: v6.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CallDetailsBean {
    'callId'?: string;
    'interactionId'?: string;
    'campaignId'?: string;
    'phone'?: string;
    'callType'?: string;
    'dispositionCode'?: string;
    'hangupCause'?: string;
    'hangupCauseCode'?: string;
    'hangupCauseCodeDescription'?: string;
    'hangupDetails'?: string;
    'hangupOnHold'?: boolean;
    'callOriginateTime'?: string;
    'callEndTime'?: string;
    'setupTime'?: string;
    'ringingTime'?: string;
    'ivrTime'?: string;
    'talkTime'?: string;
    'holdTime'?: string;
    'recordingURL'?: string;
    'callResult'?: string;
    'systemDisposition'?: string;
    'customerId'?: string;
    'leadId'?: string;
    'crmURL'?: string;
    'dateAdded'?: string;
    'timeAdded'?: string;
    'userId'?: string;
    'userName'?: string;
    'contactCenterId'?: number;
    'processId'?: number;
    'transferTo'?: string;
    'maskedPhone'?: string;
    'status'?: string;
    'dialingComments'?: string;
    'shouldDisplayMaskedPhone'?: boolean;
    'attemptTime'?: string;
    'userType'?: string;
    'crtObjectId'?: string;
    'queueName'?: string;
    'agentQueueId'?: number;
    'agentTalkTime'?: string;
    'lastReviewerName'?: string;
    'callNotes'?: string;
    'groupIds'?: Array<number>;
    'dispositionClass'?: string;
    'filterGroupSource'?: string;
    'filterSource'?: string;
    'displayPhone'?: string;
    'uniqueIdentifier'?: string;
    'groupDetail'?: string;
    'directGroupsDetail'?: string;
    'maskedNumber'?: string;
    'callScored'?: boolean;
}