// src/boilerplate/dataEngineApis/models/call-details-request-bean.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Data Engine
 * Collection of all Data Engine Requests which comprises of Reports, Stats, Monitoring and VLA Requests
 *
 * The version of the OpenAPI document: v6.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { CustomFilters } from './custom-filters';
// May contain unused imports in some cases
// @ts-ignore
import type { FilterContainer } from './filter-container';

export interface CallDetailsRequestBean {
    'searchText'?: string;
    'sortFields'?: { [key: string]: boolean; };
    'pageNumber'?: number;
    'pageSize'?: number;
    'fieldsToReturn'?: Array<string>;
    'filters'?: Array<FilterContainer>;
    'customFilters'?: CustomFilters;
}