// src/boilerplate/dataEngineApis/models/filter-container.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Data Engine
 * Collection of all Data Engine Requests which comprises of Reports, Stats, Monitoring and VLA Requests
 *
 * The version of the OpenAPI document: v6.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Contain } from './contain';
// May contain unused imports in some cases
// @ts-ignore
import type { Range } from './range';

export interface FilterContainer {
    'matchAll'?: boolean;
    'equalFilters'?: { [key: string]: string; };
    'notEqualFilters'?: { [key: string]: string; };
    'multiValueEqualFilters'?: { [key: string]: Array<string>; };
    'multiValueNotEqualFilters'?: { [key: string]: Array<string>; };
    'containFilters'?: { [key: string]: Array<Contain>; };
    'rangeFilters'?: { [key: string]: Array<Range>; };
    'filterContainerFilters'?: Array<FilterContainer>;
}